# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
host: localhost:5072
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/addSubscription:
    # binds a127 app logic to a route
    x-swagger-router-controller: payment_ctrl
    post:
      tags: ['Payment Module']
      description: Add Subscription
      summary:  Update Subscription
      # used as the method name of the controller
      operationId: addSubscription
      parameters:
        - name: body
          in: body
          description:  Update Subscription
          schema:
           $ref: '#/definitions/addSubscription'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getPaymentList:
    # binds a127 app logic to a route
    x-swagger-router-controller: payment_ctrl
    get:
      tags: ['Payment Module']
      description: Get Payment List
      summary: Get Payment List
      # used as the method name of the controller
      operationId: getPaymentList
      parameters:
        - name: page
          in: query
          description: Get Payment List
          required: false
          type: string      
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"             
...
  /api/getJobmapdata:
    # binds a127 app logic to a route
    x-swagger-router-controller: map_ctrl
    post:
      tags: ['Map Module']
      description: Get user Job List
      summary: Get user Job List
      # used as the method name of the controller
      operationId: getJobmapdata
      parameters:
        - name: body
          in: body
          description: get user By Id
          schema:
           $ref: '#/definitions/getJobmapdata'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getjobMapById:
    # binds a127 app logic to a route
    x-swagger-router-controller: map_ctrl
    post:
      tags: ['Map Module']
      description: Get user One Job 
      summary: Get user One Job
      # used as the method name of the controller
      operationId: getjobMapById
      parameters:
        - name: body
          in: body
          description: Get user One Job
          schema:
           $ref: '#/definitions/findonejob'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"           
...
definitions:
...
  addSubscription:
    type: object
    properties:
      email:
        type: string
...
  getJobmapdata:
    type: object
    properties:
      users_id:
        type: string
  findonejob:
    type: object
    properties:
      user_id:
        type: string
      job_id:
        type: string
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string